Add files via upload
in this "new" dofile I fixed most of the variables (apparently if I write them in line and not in column it works. Sometimes there were spaces after a var name that messed up the gen). Waves from 2011 to 2018 are correctly matched, the other datasets are not, I still don't know why. Have a look at it ;)